update test generator
