+++ packages/autorest.gotest/src/generator/mockTestGenerator.ts
@@ -25,10 +25,7 @@ import { Helper } from '@autorest/testmodeler/dist/src/util/helper';
 import { generateReturnsInfo, getAPIParametersSig, getClientParametersSig, getSchemaResponse } from '../util/codegenBridge';
 import { isLROOperation, isMultiRespOperation, isPageableOperation } from '../common/helpers';
 import _ = require('lodash');
-import { sortParametersByRequired } from '../common/helpers';
 export class MockTestDataRender extends BaseDataRender {
-  public clientFactoryParams: Array<Parameter>;
-
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   public skipPropertyFunc = (exampleValue: ExampleValue): boolean => {
     // skip any null value
@@ -43,14 +40,6 @@ export class MockTestDataRender extends BaseDataRender {
   };
 
   public renderData(): void {
-    let clientFactoryParams: Array<Parameter>;
-    const factoryGatherAllParamsFlag = this.context.testConfig.getValue(Config.factoryGatherAllParams)
-    if (factoryGatherAllParamsFlag) {
-      clientFactoryParams = this.getAllClientParameters();
-    } else {
-      clientFactoryParams = this.getCommonClientParameters();
-    }
-    this.clientFactoryParams = clientFactoryParams;
     const mockTest = <GoMockTestDefinitionModel>this.context.codeModel.testModel.mockTest;
     for (const exampleGroup of mockTest.exampleGroups) {
       for (const example of <Array<GoExampleModel>>exampleGroup.examples) {
@@ -155,58 +144,6 @@ export class MockTestDataRender extends BaseDataRender {
         }
       }
     }
-  }
-
-  private getCommonClientParameters(): Array<Parameter> {
-    const paramCount = new Map<string, { uses: number; param: Parameter }>();
-    let numClients = 0; // track client count since we might skip some
-    for (const group of this.context.codeModel.operationGroups) {
-      const clientName = group.language.go!.clientName;
-      // special cases: some ARM clients always don't contain any parameters (OperationsClient will be depracated in the future)
-      if (clientName.match(/^OperationsClient$/)) {
-        continue;
-      }
-      numClients++;
-      if (group.language.go!.clientParams) {
-        const clientParams = <Array<Parameter>>group.language.go!.clientParams;
-        for (const clientParam of clientParams) {
-          let entry = paramCount.get(clientParam.language.go!.name);
-          if (!entry) {
-            entry = { uses: 0, param: clientParam };
-            paramCount.set(clientParam.language.go!.name, entry);
-          }
-
-          ++entry.uses;
-        }
-      }
-    }
-    // for each param, if its usage count is equal to the
-    // number of clients, then it's common to all clients
-    const commonClientParams = new Array<Parameter>();
-    for (const entry of paramCount.values()) {
-      if (entry.uses === numClients) {
-        commonClientParams.push(entry.param);
-      }
-    }
-    commonClientParams.sort(sortParametersByRequired);
-    return commonClientParams;
-  }
-
-  private getAllClientParameters(): Array<Parameter> {
-    const allClientParams = new Array<Parameter>();
-    for (const group of this.context.codeModel.operationGroups) {
-      if (group.language.go!.clientParams) {
-        const clientParams = <Array<Parameter>>group.language.go!.clientParams;
-        for (const clientParam of clientParams) {
-          if (allClientParams.filter((cp) => cp.language.go!.name === clientParam.language.go!.name).length > 0) {
-            continue;
-          }
-          allClientParams.push(clientParam);
-        }
-      }
-    }
-    allClientParams.sort(sortParametersByRequired);
-    return allClientParams;
   }
 
   // get GO code of all parameters for one operation invoke
